Add navigation from dashboard to task board screen
