+++ feature/dashboard/src/main/java/com/jetapps/jettaskboard/drawer/DrawerHeader.kt
@@ -1,4 +1,4 @@
-package com.jetapps.jettaskboard.presentation.drawer
+package com.jetapps.jettaskboard.drawer
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Divider
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.Text
