+++ feature/dashboard/src/main/java/com/jetapps/jettaskboard/drawer/JtbDrawer.kt
@@ -1,4 +1,4 @@
-package com.jetapps.jettaskboard.presentation.drawer
+package com.jetapps.jettaskboard.drawer
 
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.layout.Column
@@ -10,8 +10,6 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Divider
 import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Home
@@ -21,13 +19,11 @@ import androidx.compose.material.icons.twotone.Settings
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.jetapps.jettaskboard.DashboardViewModel
 import com.jetapps.jettaskboard.component.NavigationDrawerItem
-import com.jetapps.jettaskboard.presentation.DashboardViewModel
 
 @Composable
 fun JtbDrawer(
